refactor: fix cycle in SignatureForm
